:sparkles: graph init yneth
Initialize the yneth subgraph.
Note that we're using the implementation contract for the init phase
so all the events can be retrieved, but we will later change the address
to be the proxy contract (0x09db87A538BD693E9d08544577d5cCfAA6373A48)
once the code is scaffolded.
```
graph init yneth yneth \
--from-contract=0x14DC3d915107DcA9ed39E29E14fbdfe4358a1346 \
--network=mainnet \
--protocol=ethereum \
--skip-git \
--contract-name=ynETH \
--index-events \
--start-block=19839557
```